terra.weather/requirements.txt
Flask==2.0.1
pymongo==3.12.0
pyowm==3.2.0
python-decouple==3.5
